+++ examples/helper.rs
